requirements.txt
faiss-cpu
sentence-transformers
gradio
